+++ SequoiaDB/engine/include/ossVer_Autogen.h
@@ -6,7 +6,7 @@
 #define SDB_ENGINE_INCLUDE_AUTOGEN_RELATIVE_PATH "../../SequoiaDB/engine/include/"
 #define SDB_ENGINE_OSSVER_FILE_RELATIVE_PATH SDB_ENGINE_INCLUDE_AUTOGEN_RELATIVE_PATH SDB_ENGINE_OSSVER_FILE
 
-#define SDB_ENGINE_RELEASE_CURRENT 34707
+#define SDB_ENGINE_RELEASE_CURRENT 34714
 #define SDB_ENGINE_BUILD_FORMAT "%04d-%02d-%02d-%02d.%02d.%02d"
 #define SDB_ENGINE_BUILD_CURRENT "0000-00-00-00.00.00"
