+++ driver/mysql/sdb_idx.h
@@ -16,8 +16,17 @@
 #ifndef SDB_IDX__H
 #define SDB_IDX__H
 
-#include "sql_class.h"  
+#include "sql_class.h"
+#include "include/client.hpp"
 
 const char *sdb_get_idx_name( KEY *key_info ) ;
+
+int build_match_obj_by_start_stop_key( uint keynr,
+                                       const uchar *key_ptr,
+                                       key_part_map keypart_map,
+                                       enum ha_rkey_function find_flag,
+                                       key_range *end_range,
+                                       TABLE *table,
+                                       bson::BSONObj &matchObj ) ;
 
 #endif
